Rename SmarHome_Copilot module
