+++ custom_components/SmartHomeCopilot/README.md
