+++ custom_components/SmartHomeCopilot/__init__.py
