+++ custom_components/SmartHomeCopilot/automations/new-entity-automation.yaml
