+++ custom_components/SmartHomeCopilot/automations/weekly-review-automation.yaml
