+++ custom_components/SmartHomeCopilot/const.py
