+++ custom_components/SmartHomeCopilot/coordinator.py
