+++ custom_components/SmartHomeCopilot/manifest.json
@@ -4,9 +4,9 @@
   "codeowners": ["@fjoelnr"],
   "config_flow": true,
   "dependencies": [],
-  "documentation": "https://github.com/fjoelnr/ai_automation_suggester",
+  "documentation": "https://github.com/fjoelnr/SmartHomeCopilot",
   "iot_class": "cloud_polling",
-  "issue_tracker": "https://github.com/fjoelnr/ai_automation_suggester/issues",
+  "issue_tracker": "https://github.com/fjoelnr/SmartHomeCopilot/issues",
   "requirements": [
     "openai>=1.0.0,<2.0.0",
     "anthropic>=0.8.0",
