+++ custom_components/SmartHomeCopilot/sensor.py
