+++ custom_components/SmartHomeCopilot/services.yaml
