+++ custom_components/SmartHomeCopilot/strings.json
