+++ custom_components/SmartHomeCopilot/translations/ca.json
