+++ custom_components/SmartHomeCopilot/translations/de.json
