+++ custom_components/SmartHomeCopilot/translations/en.json
