+++ custom_components/SmartHomeCopilot/translations/es.json
