+++ custom_components/SmartHomeCopilot/translations/it.json
