+++ custom_components/SmartHomeCopilot/translations/nl.json
