+++ custom_components/SmartHomeCopilot/translations/ru.json
