+++ custom_components/SmartHomeCopilot/translations/zh.json
