+++ custom_components/SmartHomeCopilot/www/SmartHome_Copilot/SmartHome-Copilot-card.js
